Shift world coordinates extent by 1/2 pixel width
This applies to image and labels layers, but does not modify the extent of other layer types.

Update check_layer_world_data_extent test utility to have a flag controlling whether the world
extent corresponds to coordinates. This will be set to True in tests involving Image or Labels layers.