[Experiment] Add global test timeout and slow marker to avoid regression
This set a global timeout for each test to 5s, Individual test can be
marked with a longer timeout instead of marking with with
`@pytest.mark.timeout(new_value)`, I introduce a `@slow` decorator and
marker.

1) it makes test easy to disable with "-m 'not slow'" on pytest CLI,
and
2) introduce the NAPARI_TESTING_TIMEOUT_SCALING env variable, so if you
want to double all timeouts as you are on a slower machine, you can set
it to '2'.

I expect for example to set it on macos and windows which on CI may be
quite slow.