Center data within points thumbnail (#3406)
This is the minimum amount of work I could do to fix the bug described in #1450, where the points thumbnail did not center the data as desired/expected.

The bug existed because `Points._update_thumbnail` fills out a 32x32 image regardless of the shape of the points layer's data extent. By contrast, `ImageBase._update_thumbnail` returns an image where one of the dimensions should be 32 and the other will be in [1, 32] depending on the anisotropy of the shape of the image.

The fix calculates the non-isotropic `thumbnail_shape` in `Points._update_thumbnail` in the same way that `Image._update_thumbnail` does. Padding and centering happens in `Layer.thumbnail.setter`. Using the same example in the bug description, this seems to work.

There is some opportunity for code sharing between the different implementations of `update_thumbnail`, though I haven't attempted that here yet. I also believe there are similar bugs to this one for `Vectors`, `Tracks`, and maybe `Shapes` (note that `Surface` does not have a thumbnail).

In the long term, I'm wondering if we can reuse vispy to render thumbnails (e.g. using [offscreen rendering](https://github.com/vispy/vispy/blob/main/examples/demo/gloo/offscreen.py)) to avoid maintaining different rendering pipelines in `Layer._update_thumbnail`. I think that might be an overall easier and more maintainable solution.

Based on the code added, I think this should be a safe bug fix because the coordinate/thumbnail indexing is clipped to be in bounds.

Closes #1450.