Fix off-by one error in Dims.range for non pixel-based layers (#3444)
* Add a _ranges property to LayerList to fix ViewerModel._on_layers_change

setting the ranges property depends on layer type (e.g. offset by 0.5 pixel for
Image or Label layers). This can be computed in LayerList, simplifying the
implementation of _on_layers_change in ViewerModel.

* fix: need default extent when adding empty layers

extent here matches the one in _get_extent_world

* TST: add test cases related gh-3428

verify that there is one element in the range for a single point

* Apply suggestions from code review

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>

* move nan_to_num handling into _get_min_and_max

This avoids duplication in _ranges

Co-authored-by: Nicholas Sofroniew <sofroniewn@gmail.com>
Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>