Move some hard references to weak references.
This is itself a step toward #3629, in which during the test suite we
try to make sure there is not more than one instance of the QtViewer,
VisPyCanvas and a few other things.

It seem that there are a couple of hard references between things that
can make it hard for object to be GC'd. In particular docwidgets can be
directly referenced from the main window, and QtViewer from certain
dockwidgets.