Try to track down mac test failures.
This tries to track down macs test failures which happen because of a
timeout in teardown for things that _should_ be relatively fast.

I believe i've narrowed it down to VisPyCanvases not being properly
cleaned after test, and down the rabit hole I went.

 - Make viewer a context manager to not have to closes things every
 time. Not sure we want to keep.
 - Canvas seemed to stay around too long, I haven't completely fixed
 that one yet but one referer is the QtViewer
 - QtViewer stay arround at least because of a bound method.

This does not make much changes locally, it make the timing a little bit
more consistant and less time is spent in teardown, so I hope this will
help CI.

Still draft but I'll use that to sync between my computers.