Use more weakrefs in test.
This is still related to trying to find the root cause of the mac
failures where tests teardown are taking forever, and having lingering
instances of objects like the VisPyCanvas / QtViewer.

It updates a couple of test to make sure there are no more references
to the qt viewer after each test during teardown.

This requires having one fixture return a weakref, and properly
disconnecting some closure keeping a ref on the qtviewers.