Fix screenshot test on hi dpi screen. (#3638)
The problem with data array with even rows/columns is that there is no
exact center. So (5,5) is slightly to to the lower-right of the grid.

On hi DPI/large enough screen, that means that the radius of the point
is not enough to overlap with the center of the screenshot as the point
is virtually smaller. At least that's my understanding in using
plt.imshow to look at the screenshot.

This fixes it locally on my recent M1 macbook pro, it does not affects
CI or my other laptop.

Another approach would be to compute the number of blue pixels and make
sure that they are roughly in a circular shape and central location, but
this is an easier fix.