Use a Weak Counter to track blocked events.
The Collection Counter is not weak, and was keeping around reference to
Camera._on_zoom_changed, with itself contain a reference to Camera,
down to the whole VisPyCanvas blocks, that was thus not freed.

Something along

... Viewer -> Camera ->EmitterGroup ... -> _block_counter -> _on_xxx_changed
(of Camera) -> VisPyCamera -> PanZoomCamera -> (A mess we want to
release).

Note that the event emitter usually support `(obj, 'meth')` instead of `obj.meth` for weak refs.
But that does not seem to be the case for blocker. I'm wondering if we
should try to detect when someone try to connect bound methods.