Try to close dask threads after usage.
This tries to make sure that dask thread worker is closed after each test.
This is not in it's own an issue, but from some of the mac failures I
have the impression there there are thread problems (one of the recent
failure was in IPython's history saving thread), so one of the things
I'd like to enable in pytest fixture is that the number of thread
before/after tests has not changed, unless the test if flagged
specifically.

This ensure that the above is true – at least from the dask scheduler.
We make sure the threaded scheduler is not setup.