Add support for negative labels (#3647)
* Add support for negative labels

At the Eurasia community meeting, @tibuch requested support for negative
values in the labels layer. This is useful when marking background,
objects, and sentinel "not yet proofread" values. It is easy for us to
support so in the absence of a good reason not to, I suggest here that
we do! :blush:

The only caveat is that the indexing-based colour lookup is a bit wonky.
As a reminder, for performance reasons, whenever the largest label is
not very large, we keep an array of "low discrepancy" values and we
index into that array with the labels values (#2415, and subsequent
PRs). In this case, this means that -1 indexes to the end of the array,
so if you have more than 50 labels but fewer than n-pixels labels, the
color of the -1 label will match the color of the max labels.

This situation is only temporary until we move colormapping to the
shader (#3308), so I don't think we should worry about it.

* Add support for negative labels

At the Eurasia community meeting, @tibuch requested support for negative
values in the labels layer. This is useful when marking background,
objects, and sentinel "not yet proofread" values. It is easy for us to
support so in the absence of a good reason not to, I suggest here that
we do! :blush:

The only caveat is that the indexing-based colour lookup is a bit wonky.
As a reminder, for performance reasons, whenever the largest label is
not very large, we keep an array of "low discrepancy" values and we
index into that array with the labels values (#2415, and subsequent
PRs). In this case, this means that -1 indexes to the end of the array,
so if you have more than 50 labels but fewer than n-pixels labels, the
color of the -1 label will match the color of the max labels.

This situation is only temporary until we move colormapping to the
shader (#3308), so I don't think we should worry about it.

* Add space for negative values in _get_color_lookup_func

* Trigger an all vals computation on first slice

* Add tests for negative label behaviour in slicing

* Fix test_labels.py::test_add_large_colors

* Fix remaining failing tests

Co-authored-by: Matthias Bussonnier <bussonniermatthias@gmail.com>