Better track state for starting/shutting down app.
We know make get_app and quit_app bound method on the same class which
makes it easier to track wether or not we are the one that have started
the app.

In particular this will properly close app that are not named napari.

We use quit_app in the examples to make use of that logic.