Make sure dask scheduler has been torn-down between tests.
I'm not completely sure this is a public API though,
this also explicitly requests a dask fixture to shutdown dask threads,
first thing to be explicit, and second thing as the ordering of fixture
may matter, which is harder to ensure with autouse.

While in itself dask should not be a problem to keep on, it makes it a
bit harder to track threads leaks if a test may create new threads with
dask.