Replace timeout in test by duration assertion.
There seem to be some problem when using pytest timeout. This might be
due to the fact that we use  a thread to interrupt the tests in the
middle.

Let try to keep the test going, just measure the duration and fail the
tests if they took too long.

We also remove the dependency on Pytest-timeout.

Here I'm using `time.monotonic()` (wall clock), but we can also switch to
`process_time()` (that exclude sleep-time and is thus closer to
cpu-time, it might be better in there is a lot of task switching and we
just want to make sure no computation is happening.

Mitigate #3869