Do not leak QtViewer between tests.
This is a first step toward removing leaked references during the test.
So far it make sure that nothing keeps references to the QtViewer
between tests.

We do so by turning a couple of hard references into weak ones.

They might not be the obvious ones to turn into weak ones, nor the best
ones, but _at least_ they allow collections.

We also add a flag  "--save-leaked-object-graph", that when passed to
pytest tries to save a graph (pdf) of the leaked objects, which can be
useful for debugging.