fix missing builtins in bundle (#3982)
this should fix the fact that napari's built-in plugins (including samples and writers) are not being detected in the v0.4.13 bundle. Long story short, it appears that briefcase isn't including napari's entry_points.txt along with the package metadata in the bundle (though it does for dependencies like napari-svg). This manually registers it if it hasn't been detected.