MAINT: Try to push the stack parameter down in the API.
This is an attempt to try to simplify the path_or_path in many places
that is tricky to get right. The goal would be to alway pass a list of
strings and a boolean flag as far as possible. We can't go all the way
in particular because of npe1 as this is the API plugin are expecting,
but with npe2 we should be able with a manifest revision to get a new
API for reader/writers whcih is simpler both for us and plugin writers