Push the stack keyword further down the API.
Here we push the `stack=` Argument further down the API making sure the
object we manipulate are always list of strings.

This allow to simplify even a number of utility functions that actually
were incorrectly typed as they were checking for string/Path, while only
ever strings would be given. The docs was also saying that the output
type would match the input which was technically incorrect as they could
receive a Path, and would have returned a str (or collection thereof),

The last layer we may want to push that through is npe2 and npe2 shim
for npe1, at that point all internal call would always be (list_of_str,
bool),