Try to standardise API as list of strings. (#4107)
* MAINT: Try to push the stack parameter down in the API.

This is an attempt to try to simplify the path_or_path in many places
that is tricky to get right. The goal would be to alway pass a list of
strings and a boolean flag as far as possible. We can't go all the way
in particular because of npe1 as this is the API plugin are expecting,
but with npe2 we should be able with a manifest revision to get a new
API for reader/writers whcih is simpler both for us and plugin writers

* Push the stack keyword further down the API.

Here we push the `stack=` Argument further down the API making sure the
object we manipulate are always list of strings.

This allow to simplify even a number of utility functions that actually
were incorrectly typed as they were checking for string/Path, while only
ever strings would be given. The docs was also saying that the output
type would match the input which was technically incorrect as they could
receive a Path, and would have returned a str (or collection thereof),

The last layer we may want to push that through is npe2 and npe2 shim
for npe1, at that point all internal call would always be (list_of_str,
bool),

* assert type stable

* refine

* try to fix typing