Always pass list of strings to npe2.
iter_compatible_readers takes either a string or a list of strings
we are trying to standardise on only accepting list of strings.
The behavior in passing a string or a list with 1 item is the same as
iter_compatible_readers will just check that all the items have the
same extensions and then use only the first item to find compatible
readers.

With this change we can change the api of npe2 to stop accepting strings
by only touching the npe2 repo