Attempt at viewer capturing caller.
See #4098,

I'm not a huge fan as I'm scared of leaks, but ti seem to be what people
expect.

I'm tempted to think that `update_console()` should keep a hard
reference on the objects that were given to it until the console is
opened.

I'm not sure if the behavior should be enabled by default, or now.
I've enabled it by default so that all the convenience method that
return a viewer also capture local variables.