Decrease LFS size, gif -> webm.
This should reduce quite a bit the size of git LFS and loading the
napari website.

It converts a few of the gifs (or gifs depending on how you pronounce) to
webm.

Webm can be put in img tag in most recent browser and will behave like a
video. https://caniuse.com/webm current says 95.36% of global users,
97.64% if we restrict to desktop ones.

Conversion was done using:

    ffmpeg -i {base}.gif -c vp9 -b:v 0 -crf 40 {base}.webm

With a quality of 40. You can see _some_ color quantisation by looking
well, though personally on an old and slow machine this video is nicer
than gif as it is more fluid because of the smaller size I guess ?

Other advantage, I get a play/pause/scrubber with webm, I don't get with
Gif, so it's easier to replay and find when it loops.

I also add webm to `.gitattributes` so that all webm should be tracked by
LFS automatically.

Technically I'm not sure this actually reduce the LFS payload as the old
files are likely still there, but I'm going to assume git lfs is smart
enough sometime to just pull the ones from current commit (?) and/or
that we can purge later.

I'm also not super good at git lfs, so I'd love for someone else to try
this commit manually to see if I haven't screwed up anything.

$ git lfs ls-files -s  | grep tutori |grep -v png | cut -f2 -d'-' | sort
 docs/tutorials/assets/tutorials/dask1.gif (4.3 MB)
 docs/tutorials/assets/tutorials/dask1.webm (111 KB)
 docs/tutorials/assets/tutorials/dask2.gif (13 MB)
 docs/tutorials/assets/tutorials/dask2.webm (364 KB)
 docs/tutorials/assets/tutorials/painting.gif (3.6 MB)
 docs/tutorials/assets/tutorials/painting.webm (235 KB)
 docs/tutorials/assets/tutorials/point_annotator_demo.gif (18 MB)
 docs/tutorials/assets/tutorials/point_annotator_demo.webm (105 KB)
 docs/tutorials/assets/tutorials/tracks_isbi.gif (11 MB)
 docs/tutorials/assets/tutorials/tracks_isbi.webm (1.5 MB)
 docs/tutorials/assets/tutorials/viewer_layout.jpg (141 KB)
 docs/tutorials/assets/tutorials/viewer_pan_zoom.gif (3.7 MB)
 docs/tutorials/assets/tutorials/viewer_pan_zoom.webm (221 KB)