Run PNG crush on all Pngs.
That should about decrease teh size of the Png assest by 30%,
Not only decreasing the size of git lfs, but also the size of the assets
when viewing the website, which is always good for slow connection.

```
$ git lfs ls-files  | cut -f2 -d '*' |grep png | xargs  du -c | grep total
43296	total

$ git co HEAD~1
...

$ git lfs ls-files  | cut -f2 -d '*' |grep png | xargs  du -c | grep total
64224	total
```

This was run to decrease the PNG size:

```
$ git lfs ls-files  | cut -f2 -d '*'| xargs ls -S | grep png  | xargs -L1 pngcrush -brute -ow
```

Note the unlike its sibling PR that converts gif to webm (which is a
lossy process), this should be a lossless process, at least with respect
to pixels, not sure WRT metadata.