Add NAP process for major proposals
This adds a process for documenting major decisions in the napari
library and ecosystem. I propose calling them Napari Advancement
Proposals, or NAPs. The process copies SKIPs (scikit-image), which are
modeled after NEPs (NumPy), which are themselves modeled after PEPs
(Python).

The goal of NAPs is to record the discussion around major design
decisions for the library as well as major updates to our decision
processes. @andy-sweet's design docs will be a natural fit for NAPs, as
would documentation about how working groups should operate.

As you can see, the NAP process itself is the first NAP, so let's have
some discussion around it, merge this, and then start napping! :joy: