Rewrite ellipse discretization.
There was many problem with ellipse discretization, which was losing a
lot of informations about the ellipse orientation in 3D.

This almost completely rewrite it to be more general, as well as a bit
more efficient in the generation of the triangulation indices.

I'm not sure it is more readable (but it looks way more generic).

Main difficulty in reading this code is that the shapes of objects
(transforms), depends on Wether you are in 2 or 3D, but the math is the
same.

I've tried to heavily comment it for the next modification, as I'm
pretty sure we will hit some degenerate cases.

I've also modified one of the existing example that display ellipses
in addition to points where the 4 corners would be in order to properly
visually check the rendering it correct.

See #4326 for some visualisation of incorrect rendering.

The triangle index generation at the end is a bit hard to read:

```
-    triangles = np.array([[0, i + 1, i + 2] for i in range(num_segments)])
+    triangles = (
+        np.arange(num_segments) + np.array([[0], [1], [2]])
+    ).T * np.array([0, 1, 1])
```

So I'm' happy to revert it, but it is about an order of magnitude faster
than the Python list comprehension.