Fix black line ellipse (#4312)
* Add a tolerance to fix for floating imprecision

Sometimes the difference 'd' between two vertices of equal value returned a very small value (like 8.8817842e-16), which impacted the inequality tests below, leading to 'True' where it should be 'False'. Sometimes this causes 'to_mask()' applied on an ellipse shape to return a binary mask with a black line in the middle.

Inserting a small tolerance test before that prevents this.
Another possibility would be `d = np.around(vertices[j] - vertices[i], x)` where x is the number of decimal places to round.

* Update _shapes_utils.py

Increased tolerance to increase the likelihood of working in different scenarios (for example, 1e-14 was not enough to fix [this example](https://haesleinhuepf.github.io/BioImageAnalysisNotebooks/31_graphical_user_interfaces/napari_crop.html), but 1e-12 fixes it).