Bugfix/broadcast projections by reducing number of axes (keepdims=False) (#4376)
By changing the argument keepdims from True to False projections now
have a reduced number of axes. So a 3D image with shape (a, b, c) is
projected to a 2D image with shape (b, c). When added as a layer, this
2D image will then be properly broadcast.

To account for possible transforms of the original layer, the layer
with the projection is assigned transforms from the original layer, but
with the projection axis sliced out using `set_slice`.

Commits:

* Reduce projections: keepdims=F,  slice transforms
* Fix test for keepdims = F
* use `.update()` to update the metadata
* return to just using `meta` and not layer vs. proj

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>