add is_diagonal utility and Transform property (#4370)
This adds an is_diagonal property to Transforms that makes
Affine transforms more efficient in the case when they are
known to be diagonal.

I chose to ignore any entries below a certain tolerance when
determining if a matrix is diagonal. I think in practice it is
desirable as things like applying a rotation followed by an
opposite rotation on a diagonal Affine, should ideally return the
original Affine again, but in practice will have tiny off diagonal
elements (e.g. ~1e-15) due to finite floating point precision. This
scenario is covered in one of the test cases here.

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>