add test for generate_3D_edge_meshes
Also add example script to visualized triangulated edges with various
parameters to visually check any changes that could be introduced.

In particular we see that the current method path do not properly
overlap with some combinations. That might need to be fixed later, but
will mostly be used for working on #3867.