EXPERIMENT: Batch-update context manager.
This experiment stems from the observation in #3867 that a few methods
have the `update` argument as to whether or not call
`_update_display()`, and this is generally used in case of batched
updated  to not trigger update on every call.

This is error prone, so here I'm attempting to create a reentrant
context manager that batches the updates and call `_update_display()`
only when the last context manager exists.

I'm still trying to see how that should work. In particular I'm trying
to see whether `_update_display()` should be the only method to do that.
And I'm also tempted to split `_update_display()` into to methods:
One that set a boolean flag it should be called, and one that triggers
the actual rendering on context manager exists.

This does not make the code magically faster, but it does allow some
more convenient patterns, as the context manager is reentrant.
Ina perfect world you could eagerly call multiple methods that would
trigger a redraw in a loop, and only trigger a single redraw.

I'm not sure how one would go with detecting that a
`with_batch_updates()` might be useful. I'm thinking maybe a mode where
we can measure time between subsequent update and emit a warning if it's
less that a certain amount ?

Let me know what you think of the approach.