Do not allow None when coercing encodings (#4504)
Currently, we allow `None` to be passed as a value assigned to string and color encodings, in which case it is parsed to a default value. Ideally, we want to attach encoding validation to the corresponding protocol, so that we can reduce validation boilerplate in models that have encodings. Unfortunately, Pydantic may also use `None` as a special sentinel value (see https://github.com/samuelcolvin/pydantic/issues/1624 for a similar bug), which prevents us from doing this.

There's no strong reason for us to support `None` as a value assigned to an encoding, so this PR removes that functionality. We still need to support `text` value of `None` passed from a `Layer`, but that can be done easily in `TextManager._from_layer`.

Note that this should ideally be included in v0.4.16 to prevent any usage of `string=None`. 

# References

Related to #3992