Try to upload pytest json report as artifact.
Mostly I want to investigate why tests are taking forever on CI,
like up to 45 minutes. It's hard with just printing the slowest 10
durations, as this may be dues to some test having for example many
parameters.

With this I should be able to download the artifact and analyse them.