Try to speedup test.
It seem that a large part of our slow test are spent in gc.collect.
Here we try to be a bit more lenient on how aggressive we are at
collecting garbage, and only collect all generation every 50 passes.

The youngest generation is 0, and the older 2, so by asking to colelct
up to generation 1, we should catch most of the leaks.