Add mini interactive prompt to edit string_list.json
This adds a mini TUI that ask for each non translated string whether to
ignore it or not, allowing to quickly update string_list.json.

This implements a quick getch() function that read stdin for a single
character, but will only work on unix.

This roughly looks like this in a loop.

```
    i : ignore –  add to ignored localised strings
    q : quit –  quit w/o saving
    s : save and quit

    napari/_qt/dialogs/qt_plugin_dialog.py:700 'unavailable'

                      )
                  )
    >             widget.setObjectName("unavailable")
                  widget.style().unpolish(widget)
                  widget.style().polish(widget)
                  widget.action_button.setEnabled(False)
```

Alternate/addition to #4732