MAINT: Misc mypy fixes.
It seem like since about a week mypy check have been failing everywhere.
It does not seem to be a new version of mypy, since the lastest is at
least a month old, but this shoudl fix some of the errors.

In particular mypy wants properties setter to be just next to the
getters.