Use `napari/packaging` for Conda CI (#4923)
Following NAP-2, we are migrating the CI for packaging over to napari/packaging. However, due to the required cross-repository integrations, we will still run it from napari/napari thanks to the workflow_call triggers.

This allows to reuse the remote workflow with a napari/napari token, essentially. That way artifact publishing is still kept in this repo.