MAINT: bump napari-console to 0.0.5.
It does not exists yet, but will need to be published at some point.
In particular please see napari/napari-console#19 that tries to update
napari-console

Out of interest is bumping the minimum version of Python to 3.8,
and removal of deprecated napari API so that we can remove them from
napari.