Add perfmon directory for shared configs and tools (#4898)
This adds a dot/hidden `tools/perfmon` directory to the top of the repo, which stores configs and tools associated with [performance monitoring as described on napari.org](https://napari.org/stable/howtos/perfmon.html?highlight=perfmon). Storing these in the repo makes it easier to reproduce monitoring experiments and results by standardizing configurations and tooling. In particular, I could imagine asking some napari users to use some of these tools to share measurements recorded with this tool to help us better understand napari's performance in the wild.

I'm planning to use these to help understand how running slicing in napari asynchronously affects certain measurements (e.g. the time spent in `Layer.refresh`), which is why I added a config for that here. But I could imagine other areas and configs being use in the future.

The run tool is a simple wrapper that makes it easier to run one of the named configs stored in the repo. The plot/compare tools plot the outputs of this command to generate plots like the following.

Co-authored-by: Eric Perlman <perlman@gmail.com>