Adjust the slider value to include current_size (#4951)
I am not sure how this could be implemented best, therefore I will leave it here as a request and I am more than happy to implement the change after some discussion :) I put a first draft here for a start.

When creating a Point layer the size slider is limited to be between 0 and 100.
However, in our field of cryo-em we would like to visualize sizes of up to 1000 (or more dependent on the protein) to verify the centering of the image within the point.

Therefore, I wrote a plugin to set the size of the respective points and the value of `current_size` so that new points added will have the same value. However, this means that the `current_size` value and the respective slider value are out of sync, which is very confusing for the user.

In my initial draft I propose to adjust the slider limits so that the `current_size` value is included in the range and therefore properly displayed. 


I also noticed that the way events are emitted in the points layer are kind of random!? 
Some are only emittered when `_update_properties` is `True` and  points are selected (like `current_edge_width` and `current_size`) but others are just always emitted (like `current_edge_color`). Therefore I needed to update the event manually in my test, which feels kind of wrong but maybe there is more to it :)