Start adding a CODEOWNERS file.
I suggest this after a comment from jni, as well as  pr-4932 and 4962
that I might have merged a bit overzealously without waiting for a core
napari contributor review.

I think that having assigned reviewers for part of the codebase,
and enforcing that they review/approve given changes might make ti a bit
clearer who is responsible about what.

Note that this can be made a bit easier with teams. We could have a core
team, doc team, plugins team that could be responsible for the
corresponding folders.

Current this just added Juan to review all PNGs, but that's of course
just an example.

You can find the documentation about CODEOWNERS here for more details.

https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners