Use same name in yml file and test name.
There are a number of case – like bots in #5020 that refer to the
filename, while other (Github UI) refer the name.

I never remember the Full Test <=> Comprehensive Test mapping and
end up speeding time trying to figure things out.

This sync the name of the file and the name of the tests, and I chose
Comprehensive as grepping through the code base "full", is present in
many other places, while comprehensive is more rare.