MAINT: Don't deprecate something that was never stable.
Interpolation2d is new since 0.4.17, so technically we
could can directly fail if interpolation2d is set to bilinear and move
the warning to the interpolation setter to keep backward compatibility.

This way we should never have users of interpolation2d that set an
incorrect value.

See also #5067 sibling PR