Attempt to fix benchmark failing.
Should close #5161.

I'm unsure if this is the correct way to add dependencies,
currently it seem like asv is installing with `[all]` but zarr is
technically not a dependency, just used in tests and bench.

Note that the PRs don't run the benchmarks, so this will likely need to
be merged and wait ~24h to be fixed