MAINT: Use positional only to make code a bit more idiomatic.
Up until we were ssupporting 3.7,  we had to pass args/kwargs as
non-unpacked arguments in in _make_viewer_then, as otherwise `add_method`
could be ambiguous.

Now that we support 3.8 only, we can simplify a bit _make_viewer_then.

- This avoid manually popping `viewer` from kwargs if it exists.
- Avoid making intermediate dict and tuple.

I think it is a bit easier to read, and that *args/**kwargs are a bit
better supported by linter/type inference. While we are still highly
dynamic here with the method we will call, at least we don't go through
opaque Dict[str, any].