Empty commit to debug ANACONDA_TOKEN issues
Our build actions are failing at the upload-to-anaconda.org step with an
expired token error. We have updated the tokens but get the same error
with that action. This commit attempts to upload a new tag in case GHA
is using some kind of token caching.

See Zulip discussion starting at:

https://napari.zulipchat.com/#narrow/stream/215289-release/topic/0.2E4.2E17/near/307436817