Suppress color conversion warning when converting invalid LAB coordinates (#5386)
Closes https://github.com/napari/napari/issues/5311

Our default Labels colormap is generated by sampling uniformly at random
from the unit cube. These 3-tuples are intended to be coordinates in the
CIELAB colorspace, but some of those coordinates may not be valid since
that space is not linear. We then convert those coordinates to their
corresponding RGB vectors. The scikit-image color conversion function
handles this by projecting those invalid vectors, but it also issues a
warning when doing so. This whole process approximates sampling
uniformly at random over the CIELAB colorspace and the projection adds a
little more error to the approximation, but overall it's good enough.

Originally (in #175), we vendored code from scikit-image to handle color
space conversion and removed this warning in our copy (see d1bb715).
Later, we brought in scikit-image as a dependency. Then recently (in
#5180) we removed the vendored code, which caused the warning to be
issued (#5311).

This PR specifically suppresses that warning.

It also expands on the explanatory comment describing how the min and
max values in the LAB and LUV spaces were precomputed.