Ensure four colors are in available label range
The code currently has not one but *two* hacks to make sure that nonzero
labels land in a non-transparent color. The first is that the colormap
generated has the "first" control point away from zero, by 10^(-5). The
second is that the low discrepancy image remaps labels to [1/256,
255/256] rather than to [0, 1].

These interact in that:
- the first and last colors are not selected in equal probabilities to
  the other colors.
- there is a final "phantom color" that is not reachable, with control
  points in [1-10^(-5), 1].

The fix in this case is to make the "phantom color" extra and not count
in the count of colors in the colormap. (ie if the user requests n
colors, we make n+1 because the final one never gets used.

In the future we might want to align the margins, but there is no
reliable mechanism for doing that currently.