Maint: Typing and ImportError -> ModuleNotFoundError.
Module not found is more specific than Import Error, and allows to
bubble errors when the issue is not that the module is not installed.

In the same vein it is an anti pattern to use try/except for conditional
import, as it may hide errors in Python install, using a explicit
compare with sys.version_info, can allow some tools (like pyupgrade) to
detect dead branches.