Clarify package installer interface with abstract base class, and Pip/Conda subclasses (continued) (#5124)
The UI now takes a Queue of InstallerTasks, which can be either conda or pip based commands. Right now the UI still chooses one of the two for all actions depending on whether we are running from a conda bundle or not, but in a future iteration this should depend on which action button triggered the task.

This makes an Abstract interface that the package installers in the qt plugin dialog need to implement, then refactors out all the if/then conditionals of the old Installer class into two pip/conda subclasses. Also adds (new) tests for both cases.

Co-authored-by: aganders3 <aganders3@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Talley Lambert <talley.lambert@gmail.com>
Co-authored-by: Matthias Bussonnier <bussonniermatthias@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>