Maint: Fix sporadic QtDims garbage collection failures by converting some stray references to weakrefs
As far as I can tell this is due to issue with garbage collection issue
between the worker and thread. Mostly between the garbage collection and
the worker cleanup if the last reference to the thread is set to None
and collected while the thread is still active it causes a crash.

This has two side:
1) changes to make this crash more reliably, by removing unused
   reference to the thread, and using weakrefs to make sure we free up
   things faster. Freeing things faster mean we crash more reliably.
2) actually fix the cleanup to be done after thread termination and not
   worker termination. Due to 1) we are more confident this fixes
   things.

In particular, the animation thread, has a worker that refers to the
dims via a callback/slot.

We also remove the local reference to thread, to make sure there is not
a standing, unused reference to thread, thus self._animation_thread will
be the last standing reference.

That is to say we currently have:

 - worker stop
    - delete last thread reference
 (if GC, crash)
 - Thread stops.

Now we have
 - worker stop
 - Thread stops.
    - delete last thread (and worker) reference