Remove layer ndisplay event (#5491)
# Description
This removes the private event `Layer.events._ndisplay`. This was needed
to notify the layer control widgets, some of which should only be
presented in 2D others in 3D. Now, each layer control widget stores
`ndisplay` as a public mutable property and triggers similar updates
when `ViewerModel.dims.ndisplay` changes.

There are two motivations for this change.

1. It makes `Layer._slice_dims` simpler and more clearly similar to
`Layer.refresh`.
2. We use and test against the public API of the widgets and model
rather than rely on a private event.

The main downside of this change is that the layer control widget has
some extra state (`ndisplay`) that we need to keep in sync. Initially, I
hoped we could pass through the new value of `ndisplay` through events
(rather than store it as state), but things are complex enough in
`QtImageControls` to make that difficult.

## Type of change
This is mostly a refactor with no napari public API changes.

# References
Partly motivated by #4795

# How has this been tested?
- [x] all existing tests pass with my change