Clarify layer's editable property and separate interaction with visible property (#5413)
This started off as a way to help me clarify `Layer._set_editable` and
why it's called in `Layer._slice_dims` and lots of places near
`Layer.refresh` or `Layer._update_dims` (as part of the async slicing
work). Then it snowballed a bit into a bug fix for #1346, with some
tests to prove that, which results in some refactoring and clean up on
the layer controls.

I also fixed some Qt6 compatibility issues related to `Qt.CheckState`.

- [x] Bug-fix (non-breaking change which fixes an issue)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

I think this really a bug fix, but want to flag that it does remove the
side-effect of changing `Layer.editable` when changing `Layer.visible`,
so will break anyone depending on that.

Closes #1346

- [x] added new tests to cover expected behavior
- [x] all existing tests pass with my change

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>