Add 'html-live' make action to support live reload workflows (#75)
# Description
This PR adds a `html-live` make action so we can write docs locally
without having to rebuild and serve manually. This depends on
`sphinx-autobuild`.

This is all you need:

```bash
$ make html-live
```

The incremental builds are still a bit slow, so it's not exactly "live"
but, it's indeed automated :) I've been using this in the background
while live previewing the markdown files on VS Code directly. Then, I
switch to the browser to double check.

## Type of change
- [X] Fixes or improves workflow, documentation build or deployment