Fix vispy axes labels (#5565)
# Description
This ensures that the vispy axes overlay labels are updated when the
general data of the overlay changes. One common way this happens is as a
result of a change to `Dims.ndim` which may occur when layers are added
or removed.

I added a test to cover the bug in #5536. This relies on the same
utility function I have used previously to get the vispy node scene
size, so I brought that into the general test utilities.

I also added a few smaller unit tests for the vispy overlay as sanity
checks before I found the actual issue.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# References
Closes #5536

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>