Fix inverted LUT and blending (#5487)
# Description
Followup on discussion
https://github.com/napari/napari/discussions/3914#discussioncomment-4675835.
This is an attempt at solving those issues.

This change blends "correctly" with the canvas in a translucent way, no
matter the color of the canvas, and it does so transparently to the user
(everything happens in the vispy backend). So you can actually change
opacity of every layer and have it behave as expected, even with a
non-black canvas.

Unfortunately, openGL [does not use blending parameters for the MIN and
MAX blending
functions](https://www.khronos.org/opengl/wiki/Blending#Blend_Equations).
So, if you have multiple channels in separate `minimum` blending image
layers, you have 2 options:
- the bottommost one is *not* `minimum` -> you get ugly stuff if you
change opacity to not be 1
	- the bottommost layer is `minimum` -> you cannot change opacity

As mentioned in the discussion, a way around it may be to actually
reroute the changes to `opacity` on a layer with `minimum` blending so
that they actually modify the `contrast limits`. But that's even more
magic :p

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>