Fix translation checks in CI (#5562)
# Description

The scheduled translation check on CI has been failing for 2 years. This
PR attempts to fix it by bringing everything back up to date - fixing
translations from changes to the napari codebase and changes to
pydantic.

As far as I know, the mechanics are not in place to _actually_ translate
anything (https://github.com/napari/napari/issues/2195), but this check
was put in place as a first step towards it. Keeping it up to date prior
to release is recommended.

Fixes #5019
Related (maybe resolves): #3642